requirements.txt
dash>=2.0.0
dash-extensions>=0.0.60
clean-fid
opencv-python==4.5.3.56
pandas==1.3.3
phasepack==1.5
Pillow==8.3.2
pyFFTW==0.12.0
PyYAML==5.4.1
scikit-image==0.18.3
scipy==1.7.1
staintools==2.1.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
